increasing brightness step value
for some reason, key repeat is way slower for the brightness keys, and it
wasn't always like this. I could fix it better with my own key repeat logic,
were it not for the fact that I immediately receive a KeyRelease event after
KeyPress for the brightness key events.

`xev` confirms it's only a thing with brightness keys. e.g. volume keys have
normal key repeat.