dwm started crashing too because of color emojis
Specifically, this happens when color emojis are placed into window titles. dwm
must render that title, and has problems with color emojis. This happens e.g.
when you browse a webpage with color emojis in the title (because browsers tend
to display the webpage title in the window title)

I've had the same problem with `st` too, and luckily the solution is "portable"
from st

Funny that I haven't had this problem before, even on Ubuntu 22.04. Maybe it's
a combination of things, e.g. hardware/drivers? I should probably also see if
I'm running an older dwm on the boxes that don't exhibit this problem.